Add files via upload
Welcome!
Important: This exercise is designed to test your skills and expertise in Excel. Please use a computer to complete the tasks. The data provided in these sheets are for TESTING PURPOSES ONLY.

This spreadsheet contains three tabs that will be used to assess your ability to complete sample tasks.

Below is a basic description of each tab and its contents. More detailed instructions are located on each tab.

Worksheet 1: This worksheet requires the use of basic cell formatting, conditional formatting, and the use of COUNTIF formulas.
Worksheet 2: This worksheet requires the use of PIVOT tables.
Worksheet 3: This worksheet requires the use of VLOOKUP/INDEX, LIST, and conditional formatting.